requirements.txt
datetime
json
logging
telegram.ext
random
math
requests
pymorphy2
sqlalchemy